try fix build failure
